Harden 4.x verification: skip-clean tests, zero-warning build, all_tests sync
- 4.1: 12 real-data existence gates across 9 tests converted QVERIFY->QSKIP;
  verified skip-without-data and pass-with-data; replaced CMake
  if(EXISTS hardcoded-path) guards with unconditional registration.
- 4.6: from-scratch rebuild exposed 25782 warning lines; fixed C4373 via
  const-correct Collection<T> overrides, disabled Qt-internal C4714 with
  justification, fixed 8 real diagnostics (includes, unused locals,
  int->quint32, shadowing, printf formats, getenv). Rebuild is zero-warning.
- all_tests DEPENDS was stale (8 newer tests missing, 2 removed listed);
  now verified identical to the built set. ctest 124/124 green.
- REMAINING.md 4.x status notes updated.

+++ tests/test_cellrecord.cpp
@@ -31,7 +31,7 @@ void TestCellRecord::testXclwRoundTrip()
     rec.formId = 0x2000A;
     rec.flags = 1;
     rec.cellX = 5;
-    rec.cellY = -3;
+    rec.cellY = static_cast<quint32>(-3);
     rec.owner = 0x1234;
     rec.lockLevel = 0;
     rec.cellName = QStringLiteral("Water Cell");
